fix(product): 修复商品服务中的参数传递和依赖配置问题
- 修改Feign客户端方法参数注解，将price参数从@RequestBody改为@RequestParam
- 在SpuInfoEntity实体类中添加weight字段用于存储商品重量
- 更新数据库命名空间配置从prod环境切换到dev环境
- 优化商品服务实现类中的会员价格处理逻辑，支持会员价格转换
- 调整品牌ID和分类ID的数据类型从int改为Long以支持更大数值范围
- 修复优惠券服务中SKU减少保存方法的参数接收方式
- 配置多个服务模块排除网关依赖避免冲突
- 统一依赖管理中的Web启动器配置位置

+++ gulimall-Parent/services/gulimall-product/src/main/java/com/xjz/gulimall/product/dto/SpuSaveDto.java
@@ -18,8 +18,8 @@ import java.util.List;
 public class SpuSaveDto {
     public String spuName ;
     public String spuDescription ;
-    public int catalogId ;
-    public int brandId ;
+    public Long catalogId ;
+    public Long brandId ;
     public BigDecimal weight ;
     public int publishStatus ;
     public List<String> decript ;
